Add files via upload
This is a program for an ATM machine which includes 4 options which are deposit, withdrawal, check balance and quit the app. Each option has a unique number which will allow you to choose the option you want, the program also has data validation where it will respond with an error message if you enter an invalid option.